refactor(auth, currency, user): reorganize request body types into dedicated dto.go files
- Move LoginInput, RegisterCurrencyRequest, and user-related types from their respective files into new dto.go files for better organization and separation of concerns.
- This restructuring enhances code clarity and maintainability by grouping related data types together.
- Update imports accordingly to reflect the new structure, ensuring seamless integration with the rest of the application.
