+++ webapi/auth/dto.go
@@ -1,0 +1,7 @@
+package auth
+
+// LoginInput represents the request body for user authentication.
+type LoginInput struct {
+	Identity string `json:"identity" validate:"required"`
+	Password string `json:"password" validate:"required"`
+}
